♻️ chrome.tsとutils/chrome.tsの機能を整理し、タブ関連の処理を追加
