+++ src/function/chrome/open.ts
@@ -1,29 +1,5 @@
+import { actionRuntimeContent, actionTabsContent } from "@/utils/chrome";
 import { MessageType, ActionType } from "@/types/chrome";
-
-const actionRuntimeContent = (
-  message: MessageType.OPEN_POPUP | MessageType.CLOSE_POPUP
-) => chrome.runtime.sendMessage({ type: message });
-
-const actionTabsContent = async (
-  message: MessageType.OPEN_POPUP | MessageType.CLOSE_POPUP
-) => {
-  chrome.tabs.query({ active: true, currentWindow: true }, function (tabs) {
-    const tabId = tabs[0].id;
-    if (tabId) {
-      chrome.tabs.sendMessage(tabId, { type: message });
-    }
-  });
-};
-
-const closeContent = (type: ActionType) => {
-  if (type === ActionType.runtime) {
-    return actionRuntimeContent(MessageType.CLOSE_POPUP);
-  }
-
-  if (type === ActionType.tabs) {
-    return actionTabsContent(MessageType.CLOSE_POPUP);
-  }
-};
 
 const openContent = async (type: ActionType) => {
   if (type === ActionType.runtime) {
@@ -34,5 +10,14 @@ const openContent = async (type: ActionType) => {
     return actionTabsContent(MessageType.OPEN_POPUP);
   }
 };
+const closeContent = (type: ActionType) => {
+  if (type === ActionType.runtime) {
+    return actionRuntimeContent(MessageType.CLOSE_POPUP);
+  }
 
-export { closeContent, openContent };
+  if (type === ActionType.tabs) {
+    return actionTabsContent(MessageType.CLOSE_POPUP);
+  }
+};
+
+export { openContent, closeContent };
