+++ src/types/result.ts
@@ -1,0 +1,6 @@
+export enum ResultType {
+  Tab = "Tab",
+  Bookmark = "Bookmark",
+  History = "History",
+  Google = "Google",
+}
